requirements.txt
pyrogram
tgcrypto
aiosqlite
python-dotenv
aiofiles
